removes navbar from new pages on mobile
